feat!: nullifier key (#4166)
- Rename oracle call `get_secret_key` to `get_nullifier_secret_key`.
- Add new method `request_nullifier_secret_key(account: Address)` to
`PrivateContext`, which returns a nullifier secret key for an account,
and also pushes a request for the kernel circuit to verify the keys.
- Changes to `NoteInterface`:
- Change `compute_nullifier()` to `compute_nullifier(private_context:
PrivateContext)`. If a secret is used in creating a nullifier, it must
be fetched via
`private_context.request_nullifier_secret_key(account_address)`.
- Add new method `compute_nullifier_without_context()`. This will be
used in unconstrained functions where the private context is not
available, and using an unverified nullifier keys won't affect other
users. For example, it's used in `compute_note_hash_and_nullifier` to
compute values for the user's own notes.
- Add new apis to `KeyStore`:
  - `getNullifierSecretKey(pubKey: PublicKey)`
  - `getNullifierPublicKey(pubKey: PublicKey)`
- `getSiloedNullifierSecretKey(pubKey: PublicKey, contractAddress:
AztecAddress)`