chore: make coinbase and fee_recipient inaccessible (#6375)
The AVM does not have immediate plans to support access to `coinbase`,
nor `fee_recipient`. So make access to these fields impossible from the
public interface.

This involves updating the gas token contract to "not pay fee". This is
fine because [it will be done in the base
rollup](https://github.com/AztecProtocol/aztec-packages/issues/5006).

Closes #6336.